refactor: extract tests, consolidate legacy path, fix clippy (RB-ARCH-009/010)
RB-ARCH-009: Move large scenario tests from config.rs, error.rs, and
pipeline.rs into dedicated test files under tests/ directories, reducing
production source file sizes by 60-80% while preserving full coverage.

RB-ARCH-010: Consolidate legacy memory path constant and warning message
into platforms::path_policy as the single source of truth, replacing the
hardcoded constant in session_start.rs.

Additional fixes:
- Add #[must_use] to should_process in both bootstrap modules
- Add # Errors doc sections to all public Result-returning bootstrap fns
- Fix MSRV 1.85.0 compat in dedup.rs (replace unlock() with drop)
- Fix clippy pass-by-value in memvid_store::classify_open_error
- Centralize temp-env dependency in workspace Cargo.toml (0.3.6)
- Update design doc module layout to match implementation
