+++ Cargo.toml
@@ -49,6 +49,7 @@ semver = "1"
 
 # Testing
 tempfile = "3"
+temp-env = "0.3.6"
 assert_cmd = "2"
 predicates = "3"
