+++ crates/core/src/file_guard.rs
@@ -23,16 +23,7 @@ const FORBIDDEN_PREFIXES: &[&str] = &["/dev/", "/proc/", "/sys/"];
 /// - Existing file >100 MB returns `Err(FileTooLarge)`.
 /// - Existing file within size guard returns `OpenAction::Open`.
 pub(crate) fn validate_and_open(path: &Path) -> Result<OpenAction, RustyBrainError> {
-    let path_str = path.to_string_lossy();
-    for prefix in FORBIDDEN_PREFIXES {
-        if path_str.starts_with(prefix) {
-            return Err(RustyBrainError::FileSystem {
-                code: error_codes::E_FS_PERMISSION_DENIED,
-                message: format!("system path rejected: {path_str}"),
-                source: None,
-            });
-        }
-    }
+    reject_forbidden_prefixes(path)?;
 
     if !path.exists() {
         // Create parent directories if needed
@@ -48,6 +39,44 @@ pub(crate) fn validate_and_open(path: &Path) -> Result<OpenAction, RustyBrainErr
         return Ok(OpenAction::Create);
     }
 
+    validate_existing_file(path)?;
+    Ok(OpenAction::Open)
+}
+
+/// Validate an existing memory file path without mutating the filesystem.
+///
+/// Unlike [`validate_and_open`], this function never creates parent
+/// directories and returns `E_FS_NOT_FOUND` when the file does not exist.
+pub(crate) fn validate_existing(path: &Path) -> Result<(), RustyBrainError> {
+    reject_forbidden_prefixes(path)?;
+
+    if !path.exists() {
+        return Err(RustyBrainError::FileSystem {
+            code: error_codes::E_FS_NOT_FOUND,
+            message: format!("memory file not found: {}", path.display()),
+            source: None,
+        });
+    }
+
+    validate_existing_file(path)
+}
+
+fn reject_forbidden_prefixes(path: &Path) -> Result<(), RustyBrainError> {
+    let path_str = path.to_string_lossy();
+    for prefix in FORBIDDEN_PREFIXES {
+        if path_str.starts_with(prefix) {
+            return Err(RustyBrainError::FileSystem {
+                code: error_codes::E_FS_PERMISSION_DENIED,
+                message: format!("system path rejected: {path_str}"),
+                source: None,
+            });
+        }
+    }
+    Ok(())
+}
+
+fn validate_existing_file(path: &Path) -> Result<(), RustyBrainError> {
+    let path_str = path.to_string_lossy();
     let metadata = std::fs::metadata(path).map_err(|e| RustyBrainError::FileSystem {
         code: error_codes::E_FS_IO_ERROR,
         message: format!("failed to read file metadata: {path_str}"),
@@ -73,7 +102,7 @@ pub(crate) fn validate_and_open(path: &Path) -> Result<OpenAction, RustyBrainErr
         });
     }
 
-    Ok(OpenAction::Open)
+    Ok(())
 }
 
 /// Create a timestamped backup and prune old backups.
@@ -205,6 +234,22 @@ mod tests {
         assert!(result.is_err());
         let err = result.unwrap_err();
         assert_eq!(err.code(), error_codes::E_STORAGE_FILE_TOO_LARGE);
+    }
+
+    #[test]
+    fn validate_existing_missing_file_returns_not_found_without_creating_dirs() {
+        let dir = tempfile::tempdir().unwrap();
+        let parent = dir.path().join("nested").join("deep");
+        let path = parent.join("missing.mv2");
+
+        let result = validate_existing(&path);
+        assert!(result.is_err());
+        let err = result.unwrap_err();
+        assert_eq!(err.code(), error_codes::E_FS_NOT_FOUND);
+        assert!(
+            !parent.exists(),
+            "read-only validation must not create parent directories"
+        );
     }
 
     #[test]
