+++ crates/core/tests/integration/mind_recovery.rs
@@ -78,3 +78,38 @@ fn backup_count_limited_after_multiple_corruptions() {
         backups.len()
     );
 }
+
+#[cfg(unix)]
+#[test]
+fn permission_errors_do_not_trigger_recovery_backup() {
+    use std::os::unix::fs::PermissionsExt;
+
+    let dir = tempfile::tempdir().unwrap();
+    let path = dir.path().join("permission-denied.mv2");
+    std::fs::write(&path, b"valid content irrelevant").unwrap();
+    std::fs::set_permissions(&path, std::fs::Permissions::from_mode(0o000)).unwrap();
+
+    let config = types::MindConfig {
+        memory_path: path.clone(),
+        ..types::MindConfig::default()
+    };
+
+    let result = Mind::open(config);
+    assert!(
+        result.is_err(),
+        "permission-denied open should fail without recovery"
+    );
+
+    let backups: Vec<_> = std::fs::read_dir(dir.path())
+        .unwrap()
+        .filter_map(|e| e.ok())
+        .filter(|e| e.file_name().to_string_lossy().contains(".backup-"))
+        .collect();
+    assert!(
+        backups.is_empty(),
+        "non-corruption open failures must not create recovery backups"
+    );
+
+    // Restore permissions so tempfile cleanup can remove the file.
+    let _ = std::fs::set_permissions(&path, std::fs::Permissions::from_mode(0o600));
+}
