+++ crates/hooks/src/session_start.rs
@@ -1,16 +1,16 @@
 use std::path::Path;
 
+use crate::bootstrap;
 use crate::context::format_system_message;
 use crate::error::HookError;
 use types::hooks::{HookInput, HookOutput};
-use types::{MindConfig, ProjectContext};
-
-const LEGACY_MEMORY_PATH: &str = ".claude/mind.mv2";
 
 /// Handle the session-start hook event.
 ///
-/// Detects the platform, resolves project identity and memory path,
-/// opens or creates the Mind, fetches context, and returns a system message.
+/// Runs the event through the adapter + pipeline for contract validation and
+/// identity resolution. If the pipeline skips the event, returns a default
+/// (no-op) output. Otherwise opens the Mind, fetches context, and returns
+/// a system message.
 ///
 /// # Errors
 ///
@@ -18,49 +18,24 @@ const LEGACY_MEMORY_PATH: &str = ".claude/mind.mv2";
 pub fn handle_session_start(input: &HookInput) -> Result<HookOutput, HookError> {
     let cwd = Path::new(&input.cwd);
 
-    // Detect platform
-    let platform_name = platforms::detect_platform(input);
+    if !bootstrap::should_process(input, "session_start") {
+        return Ok(HookOutput::default());
+    }
 
-    // Resolve project identity
-    let context = ProjectContext {
-        platform_project_id: None,
-        canonical_path: None,
-        cwd: Some(input.cwd.clone()),
-    };
-    let _identity = platforms::resolve_project_identity(&context);
-
-    // Resolve memory path
-    let resolved = platforms::resolve_memory_path(cwd, &platform_name, false).map_err(|e| {
-        HookError::Platform {
-            message: format!("Failed to resolve memory path: {e}"),
-        }
-    })?;
-
-    // Build MindConfig
-    let config = MindConfig {
-        memory_path: resolved.path.clone(),
-        ..MindConfig::default()
-    };
-
-    // Open Mind (creates new .mv2 if missing)
-    let mind = rusty_brain_core::mind::Mind::open(config)?;
+    let memory_path = bootstrap::resolve_memory_path(input, cwd)?;
+    let mind = bootstrap::open_mind(input, cwd)?;
 
     // Get context and stats
     let ctx = mind.get_context(None)?;
     let stats = mind.stats()?;
 
     // Format system message
-    let mut message = format_system_message(&ctx, &stats, &resolved.path);
+    let mut message = format_system_message(&ctx, &stats, &memory_path);
 
-    // Check for legacy memory path
-    let legacy_path = cwd.join(LEGACY_MEMORY_PATH);
+    // Check for legacy memory path (constant and message owned by path_policy)
+    let legacy_path = cwd.join(platforms::LEGACY_CLAUDE_MEMORY_PATH);
     if legacy_path.exists() {
-        use std::fmt::Write;
-        let _ = write!(
-            message,
-            "\n**Note:** Legacy memory file detected at `{LEGACY_MEMORY_PATH}`. \
-             Consider migrating to `.agent-brain/mind.mv2`.\n"
-        );
+        message.push_str(&platforms::format_legacy_path_warning(&memory_path));
     }
 
     Ok(HookOutput {
