+++ crates/platforms/src/lib.rs
@@ -27,6 +27,9 @@ pub use adapters::{claude_adapter, opencode_adapter};
 pub use contract::{SUPPORTED_CONTRACT_MAJOR, validate_contract};
 pub use detection::detect_platform;
 pub use identity::resolve_project_identity;
-pub use path_policy::{PathMode, ResolvedMemoryPath, resolve_memory_path};
+pub use path_policy::{
+    LEGACY_CLAUDE_MEMORY_PATH, PathMode, ResolvedMemoryPath, format_legacy_path_warning,
+    resolve_memory_path,
+};
 pub use pipeline::{EventPipeline, PipelineResult};
 pub use registry::AdapterRegistry;
