+++ crates/types/Cargo.toml
@@ -17,7 +17,7 @@ uuid = { workspace = true }
 ulid = { workspace = true }
 
 [dev-dependencies]
-temp-env = "0.3"
+temp-env = { workspace = true }
 
 [lints]
 workspace = true
