+++ docs/plans/2026-03-04-arch-006-007-008-design.md
@@ -1,0 +1,79 @@
+# Design: RB-ARCH-006, RB-ARCH-007, RB-ARCH-008
+
+**Date**: 2026-03-04
+**Tickets**: RB-ARCH-006 (P1), RB-ARCH-007 (P2), RB-ARCH-008 (P2)
+**Phase**: Phase 3 — Scalability and Maintainability
+
+---
+
+## RB-ARCH-006: Cap Over-Fetch in Filtered CLI Queries
+
+### Problem
+When a type filter is active, `run_find` and `run_timeline` use `usize::MAX` as the backend fetch limit, loading every frame into memory before post-filtering. For large stores this is wasteful.
+
+### Solution
+Replace `usize::MAX` with `limit * 10` (capped at a reasonable ceiling). This provides enough headroom for type diversity while bounding memory usage.
+
+### Changes
+- `crates/cli/src/commands.rs`: 2 lines changed (one in `run_find`, one in `run_timeline`)
+
+### Acceptance
+- No `stats()` call in filtered query flow (already true)
+- Over-fetch bounded to `limit * 10` instead of `usize::MAX`
+- All existing tests pass
+
+---
+
+## RB-ARCH-007: Decompose Mind Into Focused Modules
+
+### Problem
+`crates/core/src/mind.rs` (747 lines) mixes lifecycle/recovery, locking, read API, write API, stats, and helper types. This exceeds the 400-line guideline and makes targeted changes risky.
+
+### Solution
+Convert `mind.rs` into a `mind/` directory with focused submodules. The public API surface remains identical.
+
+### Module Layout
+
+```
+crates/core/src/mind/
+  mod.rs          — Mind struct, accessors, TimelineEntry, MemorySearchResult, re-exports
+  lifecycle.rs    — open, open_read_only, open_with_backend, recovery, permissions
+  locking.rs      — with_lock, ReentrantLockGuard, IN_FILE_LOCK
+  read.rs         — search, ask, timeline, get_context, parse_search_hit
+  write.rs        — remember, remember_unlocked, save_session_summary
+  stats.rs        — stats with caching
+```
+
+### Key Constraints
+- `pub mod mind` in `lib.rs` stays unchanged
+- `use rusty_brain_core::mind::Mind` still works
+- Tests stay in `mind_tests.rs` (referenced via `#[cfg(test)] #[path]`)
+- All `impl Mind` blocks use the same struct from `mod.rs`
+- `pub(crate)` visibility preserved for internal helpers
+
+### Acceptance
+- API behavior preserved: all integration tests unchanged/green
+- No new public types or functions
+- Each module < 200 lines
+
+---
+
+## RB-ARCH-008: Remove get_mind/reset_mind Singleton
+
+### Problem
+`get_mind`/`reset_mind` in `crates/core/src/lib.rs` have zero runtime callers. Both hooks and opencode explicitly use `Mind::open()` per invocation per spec guidance. The singleton adds global state complexity for no consumers.
+
+### Decision
+Remove the singleton API entirely.
+
+### Changes
+- Remove `get_mind`, `reset_mind`, `MIND_INSTANCE` from `crates/core/src/lib.rs`
+- Remove the `singleton_get_and_reset` test
+- Update module doc comment to remove singleton references
+- Remove `Arc` and `Mutex` imports that become unused
+
+### Acceptance
+- No compilation errors
+- No runtime callers broken (verified: zero exist outside the test itself)
+- `cargo test --workspace` passes
+- `cargo clippy --workspace -- -D warnings` passes
